refactor: Split enums into files, add test
