+++ specs/destination_test.go
@@ -10,22 +10,6 @@ type testDestinationSpec struct {
 	ConnectionString string `json:"connection_string"`
 }
 
-func TestWriteModeFromString(t *testing.T) {
-	var writeMode WriteMode
-	if err := writeMode.UnmarshalJSON([]byte(`"append"`)); err != nil {
-		t.Fatal(err)
-	}
-	if writeMode != WriteModeAppend {
-		t.Fatalf("expected WriteModeAppend, got %v", writeMode)
-	}
-	if err := writeMode.UnmarshalJSON([]byte(`"overwrite"`)); err != nil {
-		t.Fatal(err)
-	}
-	if writeMode != WriteModeOverwrite {
-		t.Fatalf("expected WriteModeOverwrite, got %v", writeMode)
-	}
-}
-
 func TestDestinationSpecUnmarshalSpec(t *testing.T) {
 	destination := Destination{
 		Spec: map[string]any{
@@ -211,17 +195,5 @@ func TestDestinationUnmarshalSpecValidate(t *testing.T) {
 				t.Fatalf("expected:\n%v\ngot:\n%v\n", tc.destination, destination)
 			}
 		})
-	}
-}
-
-func TestWriteMode(t *testing.T) {
-	for _, writeModeStr := range writeModeStrings {
-		writeMode, err := WriteModeFromString(writeModeStr)
-		if err != nil {
-			t.Fatal(err)
-		}
-		if writeModeStr != writeMode.String() {
-			t.Fatalf("expected:%s got:%s", writeModeStr, writeMode.String())
-		}
 	}
 }
