+++ specs/write_mode.go
@@ -1,0 +1,53 @@
+package specs
+
+import (
+	"bytes"
+	"encoding/json"
+	"fmt"
+)
+
+type WriteMode int
+
+const (
+	WriteModeOverwriteDeleteStale WriteMode = iota
+	WriteModeOverwrite
+	WriteModeAppend
+)
+
+var (
+	writeModeStrings = []string{"overwrite-delete-stale", "overwrite", "append"}
+)
+
+func (m WriteMode) String() string {
+	return writeModeStrings[m]
+}
+
+func (m WriteMode) MarshalJSON() ([]byte, error) {
+	buffer := bytes.NewBufferString(`"`)
+	buffer.WriteString(m.String())
+	buffer.WriteString(`"`)
+	return buffer.Bytes(), nil
+}
+
+func (m *WriteMode) UnmarshalJSON(data []byte) (err error) {
+	var writeMode string
+	if err := json.Unmarshal(data, &writeMode); err != nil {
+		return err
+	}
+	if *m, err = WriteModeFromString(writeMode); err != nil {
+		return err
+	}
+	return nil
+}
+
+func WriteModeFromString(s string) (WriteMode, error) {
+	switch s {
+	case "append":
+		return WriteModeAppend, nil
+	case "overwrite":
+		return WriteModeOverwrite, nil
+	case "overwrite-delete-stale":
+		return WriteModeOverwriteDeleteStale, nil
+	}
+	return 0, fmt.Errorf("invalid write mode: %s", s)
+}
